Move `LegacyTokenizer` to `tokenizers.legacy`
