+++ curated_transformers/tokenizers/legacy/bbpe_tokenizer.py
@@ -3,7 +3,8 @@ from typing import Dict, Iterable, List, Optional, Tuple, TypeVar
 from curated_tokenizers import ByteBPEProcessor
 
 from ..chunks import MergedInputChunks, MergedSpecialPieceChunk
-from ..tokenizer import LegacyTokenizer, PiecesWithIds
+from ..tokenizer import PiecesWithIds
+from .legacy_tokenizer import LegacyTokenizer
 
 # Only provided as typing.Self in Python 3.11+.
 Self = TypeVar("Self", bound="ByteBPETokenizer")
