+++ curated_transformers/tokenizers/legacy/camembert_tokenizer.py
@@ -5,7 +5,7 @@ from curated_tokenizers import SentencePieceProcessor
 
 from curated_transformers.tokenizers.hf_hub import LegacyFromHFHub
 
-from ..tokenizer import AddBosEosPreEncoder
+from .legacy_tokenizer import AddBosEosPreEncoder
 from ._fairseq import FAIRSEQ_PIECE_IDS, FairSeqPostEncoder, FairSeqPreDecoder
 from .sentencepiece_tokenizer import SentencePieceTokenizer
