+++ curated_transformers/tokenizers/legacy/llama_tokenizer.py
@@ -4,8 +4,8 @@ from typing import Any, Dict, Optional, Type, TypeVar
 from curated_tokenizers import SentencePieceProcessor
 
 from ..hf_hub import LegacyFromHFHub
+from .legacy_tokenizer import AddBosEosPreEncoder
 from .sentencepiece_tokenizer import SentencePieceTokenizer
-from ..tokenizer import AddBosEosPreEncoder
 
 # Only provided as typing.Self in Python 3.11+.
 Self = TypeVar("Self", bound="LLaMATokenizer")
