+++ curated_transformers/tokenizers/legacy/xlmr_tokenizer.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, Optional, Type, TypeVar
 from curated_tokenizers import SentencePieceProcessor
 
 from ..hf_hub import LegacyFromHFHub
-from ..tokenizer import AddBosEosPreEncoder
+from .legacy_tokenizer import AddBosEosPreEncoder
 from ._fairseq import FAIRSEQ_PIECE_IDS, FairSeqPostEncoder, FairSeqPreDecoder
 from .sentencepiece_tokenizer import SentencePieceTokenizer
